# Remove `conj` method