The `conj` method is defined for `AbstractArray` in `Base`, and this seems not necessary in StaticArrays.jl.
https://github.com/JuliaLang/julia/blob/master/base/abstractarraymath.jl#L120