@fredrikekre 2021-11-02
You probably meant to link to https://github.com/JuliaLang/julia/blob/c646b5d4893e45a413705d3fea0a7227c4fe7aae/base/abstractarraymath.jl#L145

It takes quite a different code path, have you measured anything?