@hyrodium 2021-11-02
I have checked some benchmarks such as the following, and got some performance improvements.

## Before this PR
```julia
julia> using StaticArrays, BenchmarkTools

julia> M = @SMatrix rand(10,10)
10×10 SMatrix{10, 10, Float64, 100} with indices SOneTo(10)×SOneTo(10):
 0.278087   0.532613  0.700905     0.6797    0.1741     0.482916   0.559134  0.416935   0.583389    0.738919
 0.112061   0.680568  0.00452278   0.629453  0.565331   0.899292   0.506236  0.858908   0.6343      0.758504
 0.265049   0.981814  0.865431     0.910921  0.0365868  0.308681   0.392093  0.045781   0.831745    0.156426
 0.990246   0.573634  0.000770404  0.506469  0.236795   0.0497442  0.183051  0.435251   0.604946    0.438511
 0.0147305  0.686109  0.403095     0.467053  0.230596   0.655527   0.789942  0.804046   0.802971    0.0322616
 0.920373   0.534842  0.359124     0.701889  0.0221901  0.116204   0.633207  0.859563   0.887927    0.154059
 0.214432   0.230608  0.190767     0.931815  0.80687    0.183066   0.894801  0.799044   0.991004    0.658214
 0.618901   0.61529   0.676653     0.905544  0.39745    0.448599   0.749681  0.661116   0.00369557  0.170434
 0.623915   0.772886  0.275751     0.522046  0.537816   0.948463   0.138827  0.0578676  0.141086    0.188904
 0.766867   0.940663  0.454173     0.723323  0.903619   0.937994   0.777659  0.497007   0.883013    0.158455

julia> @benchmark conj(M)

BenchmarkTools.Trial: 10000 samples with 972 evaluations.
 Range (min … max):   72.813 ns … 765.871 ns  ┊ GC (min … max): 0.00% … 68.04%
 Time  (median):     103.034 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   113.903 ns ±  65.503 ns  ┊ GC (mean ± σ):  7.28% ± 10.61%

    ▅█▇▅▃▂▁                                                     ▂
  ▅▆████████▆▅▄▁▁▁▁▃▁▁▅▅▅▅▄▃▁▁▄▆▆▅▄▁▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅█▇▇ █
  72.8 ns       Histogram: log(frequency) by time        549 ns <

 Memory estimate: 816 bytes, allocs estimate: 1.

julia> M = @MMatrix rand(10,10)
10×10 MMatrix{10, 10, Float64, 100} with indices SOneTo(10)×SOneTo(10):
 0.247761  0.319253  0.335513   0.31124     0.549472   0.218886  0.591535  0.933736   0.336211   0.13741
 0.279126  0.424842  0.144185   0.299496    0.877419   0.342943  0.792121  0.970909   0.602586   0.716671
 0.916509  0.191672  0.648459   0.00845573  0.309583   0.592104  0.425679  0.703165   0.368872   0.594109
 0.637092  0.410949  0.0461443  0.263881    0.49783    0.304925  0.712598  0.715469   0.124621   0.688993
 0.176864  0.514431  0.9201     0.976133    0.955894   0.433826  0.930658  0.766981   0.362407   0.882564
 0.580898  0.530565  0.679387   0.682978    0.0694423  0.996007  0.662798  0.30554    0.65317    0.0269546
 0.855626  0.472388  0.149302   0.110612    0.335256   0.693506  0.547056  0.138001   0.214018   0.10246
 0.23437   0.701765  0.196096   0.593873    0.71138    0.728251  0.476363  0.444303   0.0866761  0.642449
 0.735992  0.600846  0.923802   0.167488    0.647316   0.155028  0.579474  0.590639   0.755423   0.739292
 0.398773  0.925451  0.720805   0.65011     0.395011   0.273522  0.285149  0.0849117  0.499625   0.996491

julia> @benchmark conj(M)
BenchmarkTools.Trial: 10000 samples with 988 evaluations.
 Range (min … max):   49.922 ns … 803.401 ns  ┊ GC (min … max): 0.00% … 84.69%
 Time  (median):      95.290 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   106.538 ns ±  76.779 ns  ┊ GC (mean ± σ):  9.43% ± 11.18%

     ▅█▄▂                                                       ▁
  ▅▄▅████▇▅▄▁▃▁▁▁▁▅▄▆▇▅▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▇█ █
  49.9 ns       Histogram: log(frequency) by time        666 ns <

 Memory estimate: 816 bytes, allocs estimate: 1.
```

## After this PR
```julia
julia> using StaticArrays, BenchmarkTools

julia> M = @SMatrix rand(10,10)
10×10 SMatrix{10, 10, Float64, 100} with indices SOneTo(10)×SOneTo(10):
 0.395866  0.984503  0.0355468  0.477684  0.379548  0.0073546  0.616719   0.985038    0.652924  0.0153167
 0.669694  0.693874  0.249755   0.502189  0.39266   0.792999   0.0691833  0.606156    0.230594  0.498549
 0.794368  0.627234  0.629443   0.82231   0.85233   0.381944   0.666228   0.00887486  0.769741  0.113535
 0.469119  0.486845  0.80488    0.213074  0.823996  0.625862   0.504831   0.64473     0.910617  0.998263
 0.527923  0.638194  0.0216183  0.909469  0.758754  0.0268722  0.899368   0.184593    0.124699  0.241174
 0.104969  0.416401  0.7006     0.611044  0.950759  0.532786   0.824546   0.825787    0.981826  0.889252
 0.545392  0.283333  0.149094   0.191652  0.635965  0.869245   0.632605   0.561928    0.824314  0.264521
 0.82772   0.506062  0.77622    0.940731  0.372452  0.693001   0.23488    0.866791    0.840111  0.266015
 0.651002  0.472906  0.67735    0.505669  0.136804  0.348819   0.609913   0.65026     0.520197  0.802687
 0.323443  0.281224  0.477948   0.447889  0.459099  0.0955617  0.580284   0.920368    0.85748   0.759464

julia> @benchmark conj(M)
BenchmarkTools.Trial: 10000 samples with 995 evaluations.
 Range (min … max):  26.069 ns … 50.789 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     26.755 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   26.913 ns ±  0.964 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

       █                                                       
  ▃▁▂▂▂█▅▃█▃▄▃▆▃▇▄▅▃▃▃▃▃▃▃▂▅▂▂▃▄▄▂▃▂▃▂▂▂▂▂▂▁▂▂▂▁▂▂▂▂▂▂▂▂▂▂▂▂▂ ▃
  26.1 ns         Histogram: frequency by time        29.3 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

julia> M = @MMatrix rand(10,10)
10×10 MMatrix{10, 10, Float64, 100} with indices SOneTo(10)×SOneTo(10):
 0.8378     0.801518  0.676544    0.750425   0.290985  0.762238  0.334579   0.874218  0.95687    0.139948
 0.473978   0.549372  0.686047    0.946951   0.230506  0.114907  0.52121    0.323596  0.603784   0.403532
 0.15215    0.482342  0.550598    0.848107   0.93614   0.551035  0.996267   0.601266  0.749325   0.406892
 0.0860765  0.750351  0.854076    0.246915   0.702297  0.279293  0.0284717  0.699988  0.847526   0.718802
 0.492089   0.847479  0.00690432  0.0130087  0.759342  0.596706  0.0890115  0.51883   0.957287   0.570171
 0.576063   0.59831   0.0717801   0.909979   0.183003  0.426615  0.894805   0.503987  0.43252    0.385912
 0.27759    0.132377  0.134263    0.892036   0.664279  0.274565  0.714938   0.498226  0.999056   0.57042
 0.131551   0.542445  0.568774    0.380064   0.198061  0.284786  0.502495   0.417818  0.779565   0.891566
 0.0421118  0.361299  0.746001    0.685135   0.203758  0.583721  0.45083    0.563815  0.0405225  0.606448
 0.163135   0.114762  0.0722628   0.993868   0.53986   0.441421  0.855709   0.815099  0.125015   0.633142

julia> @benchmark conj(M)
BenchmarkTools.Trial: 10000 samples with 999 evaluations.
 Range (min … max):  9.226 ns … 19.456 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     9.377 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.495 ns ±  0.518 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

     ▂▃ █    ▂            ▁  ▃                           ▃ ▂ ▁
  ▄▆▅██▄█▇▁▃▁█▁▁▁▁▁▁▁▁▁▁▁▃█▅██▁▁▁▁▁▃▁▁▁▁▁▁▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁█▁█ █
  9.23 ns      Histogram: log(frequency) by time     10.6 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```